output part of the logic from main.js in create-film-cards.js
